# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Status:
    enum:
    - pending
    - done
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusDone
  models.Todo:
    properties:
      description:
        type: string
      status:
        $ref: '#/definitions/models.Status'
      title:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Sysware Test Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sysware Test Services API
  version: "1.0"
paths:
...
  /todos:
    get:
      consumes:
      - application/json
      description: Get all todos with pagination and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_by
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Search column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all todos
      tags:
      - Todo
    post:
      consumes:
      - application/json
      description: Create a new todo
      parameters:
      - description: Todo Data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new todo
      tags:
      - Todo
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete any todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a todo (Admin)
      tags:
      - Todo
    get:
      consumes:
      - application/json
      description: Get a todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a todo by ID
      tags:
      - Todo
    patch:
      consumes:
      - application/json
      description: Update any todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo Data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a todo (Admin)
      tags:
      - Todo
  /todos/me:
    get:
      consumes:
      - application/json
      description: Get todos for the authenticated user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_by
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Search column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get my todos
      tags:
      - Todo
    post:
      consumes:
      - application/json
      description: Create a new todo for the authenticated user
      parameters:
      - description: Todo Data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new todo for me
      tags:
      - Todo
  /todos/me/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo for the authenticated user
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete my todo
      tags:
      - Todo
    get:
      consumes:
      - application/json
      description: Get a specific todo for the authenticated user
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get my todo by ID
      tags:
      - Todo
    patch:
      consumes:
      - application/json
      description: Update a todo for the authenticated user
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo Data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update my todo
      tags:
      - Todo
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"